fix(page pagination): add timestamp cursor based pagination for tweets feed

+++ src/main/java/com/xclone/xclone/domain/bookmark/BookmarkRepository.java
@@ -25,7 +25,7 @@ public interface BookmarkRepository extends JpaRepository<Bookmark, Integer> {
 """)
     List<Integer> findPaginatedBookmarkedPostIds(
             @Param("userId") int userId,
-            @Param("cursor") int cursor,
+            @Param("cursor") long cursor,
             Pageable pageable
     );
 }
